backend/requirements.txt
# =====================================
# GAUR Backend Requirements
# Python 3.11+ required
# =====================================

# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Pydantic and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0

# Web Scraping
telethon==1.34.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/ML Models
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.0
torchvision==0.16.0
openai>=1.0.0  # GPT-4 Vision API

# Computer Vision and OCR
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10

# NLP and Language Detection
langdetect==1.0.9
nltk==3.8.1

# Machine Learning Utilities
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Embeddings and Similarity Search
faiss-cpu==1.7.4

# Background Tasks and Scheduling
apscheduler==3.10.4
celery==5.3.4

# PDF Generation for Reports
reportlab==4.0.7

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
ipython==8.18.1
